Show escaped strings correctly in log/report.
Update issue 1164
Fixed showing nowadays escaped strings correctly in log and report.

This issue is now otherwise done, but escaping strings with &xxx; breaks showing combined tags in report in special cases. This is because & is considered `AND` in those patterns. I'll leave this issue open until that is fixed.