Fixed showing combined tags in report when any of the combined tags contains <, >, or &.
This was broken when all strings written to JS model were escaped. A fix is to always use `AND` instead of `&` in combined patterns in JS. That's probably better also visually because the other special boolean operator we use is `NOT` and we may get `OR` in the future.

Update issue 1164
Status: Done
The problem with combined tags with <, >, and & is now fixed. I consider this issue done.