Show elepsed in log/report stats always as hh:mm:ss.
Update issue 1194
After a little discussion within the team, we decided to show elapsed
always as hh:mm:ss. A drawback is that you don't see millis, but we
don't expect users to be interested in them too often. The main
benifit is that the format is always the same and it is also the same
as used elsewhere. Another benefit is that always using the same
format makes it easier to implement sorting of the table.