Separete column for elaped time in report's test details table.
Update issue 1206
Status: Started
Separate column for elaped added. Decided to add end time with start time to keep row hight same as earlier.

Currently elapsed times are sorted first so that fastests tests are first. In stats elapsed times are sorted so that longest are first and that's definitely better approach here too. Unfortunately the jQuery tablesorter version we use doesn't seem support changing the initial sort order of an individual column. I'll see could that be done with a custom parser or should we upgrade tablesorter to a forked version.