XML lib: New approach for handling namespaces.
Update issue 1241
It turned out that handling prefixed namespaces using the same approach as default namespaces was somewhat complicated and end results weren't that nice. Decided to change namespace handling to strip all {namespaces} from tag names and isntead moved namespace declations to elements' xmlns attribute as the default namespace.

This approach is easier to use than the earlier and works as well as it with default namespaces. With non-default (prefixed) namespaces everything works nicely too except that those prefixes are lost. Importantly saved XML will be nevertheless be semantically same as original.

One limitation of this solution is that possible attribute namespaces are not handled at all. With them stripping {namespaces} isn't possible. Luckily namespaces are very rarely used with attributes but this limitation nevertheless needs to be documented.