Hopefully fixed Chrome sometimes hanging when opening log when tests are run like --LogLevel TRACE:INFO.
Update issue 1249
Status: Pending
Owner: pekka.klarck
Cc: mikko.korpela
We weren't able to reproduce the problem, but hopefully we were anyway able to fix it based on the description and looking at the code.

We simply changed the initial call to set the log level to use setMessageVisibility directly instead of logLevelSelected. This avoids scrolling altogether when setting the level initially, which is totally fine on the first time. The idea of the scolling in logLevelSelected is to keep same view visible after altering the level via the dropdown widget.

It would be great if you Tony could test does the fix work for you.