Fixed teardown with keywords containing embedded arguments
Update issue 1277
Status: Done
Labels: Priority-High
The reason this didn't work was that teardown of keywords with embedded arguments were uncondionally set to None. Pretty bad oversight while this functionality has been implemented.

I also noticed that the code related to handling embedded arguments was unnecessary complicated. I'll refactor that later but that's not part of this issue.