Argparser: Support any number of spaces as option/value separator in argument files.
Update issue 1281
Summary: Argument files should support '=' and any number of spaces as separator between options and values
Decided to change functionality so that any number of spaces can be used as a separator. This makes rows below equivalent:

  --name Foo
  --name=Foo
  --name       Foo

and allows this kind of argument files:

  --name      My Name
  --doc       My doc
  --include   tag

Notice that spaces are not ignored around '=' when it is used as a separator. Thus the following are now equivalent:

  --name=Name
  --name= Name