URL matching: Reverted to old algorithm with small enhancements.
Update issue 1347
Status: WontFix
Owner: -pekka.klarck
Labels: -Target-2.7.7

The new URL matching algorithm turned out to be a bit greedy. Matching
URLs in strings like `Foo:http://bar` caused problems with --tagstatlink
option (build broken) and could have caused problems also elsewhere. It
would have been better to change HTML escaping so that URL replacing can
be disabled altogether, but that's a bigger change and this starts to
get too risky to be done in a minor release.

In the end decided to revert to the old URL matching algorithm. Did some
enhancements to it, though, to make most of the newly added tests pass.
The only test that needed to be removed was handling pipes.