Fixed validating kwargs in --dryrun mode.
Update issue 1511
Status: Done
Owner: pekka.klarck
It turned out that kwargs weren't handled at all in dry-run mode. This caused
false positives in cases like in this issue, but it also caused false
negatives in some cases.

Luckily fixing the problem wasn't too complicated, because it was possible to
use the same code for resolving/validating arguments in dry-run than in normal
run. In the end more code was removed than added.