BuiltIn.Convert To Integer
Update issue 1581
Little cleanup to code and tests. Good enough for me.

It's somewhat stupid to use RuntimeError for communication here, but
that's not easy to change because _convert_to_integer uses it. The
original motivation for that has been that error messages got from
RuntimeError's don't get `RuntimeError: ` prefix similarly as you get
e.g. `ValueError: `. The best solution would probably be using a custom
exception with this prefix disabled, but that's too much work now.