Remote: Support continuable and fatal errors.
Update issue 1628
Owner: pekka.klarck
Status: Started
Implemented Remote library side of the new functionality.

Final decision was to add new optional 'continuable' and 'fatal' keys to the
result dictionary. If these keys are present, and the value is True, the
situation will be reported to Robot accordingly. These values are only used if
the keyword fails, and if both values are set then 'fatal' has precedence.